requirements.txt
Flask
fpdf
Pillow
python-pptx
gunicorn

